♻️Unfold a few functions

+++ Source/spells.h
@@ -15,7 +15,7 @@ bool CheckSpell(int id, spell_id sn, spell_type st, bool manaonly);
 void EnsureValidReadiedSpell(PlayerStruct &player);
 void CastSpell(int id, int spl, int sx, int sy, int dx, int dy, int spllvl);
 void DoResurrect(int pnum, int rid);
-void DoHealOther(int pnum, int rid);
+void DoHealOther(int pnum, uint16_t rid);
 int GetSpellBookLevel(spell_id s);
 int GetSpellStaffLevel(spell_id s);
